30% Implemented Strategy for enemy position generation
